Moving away from pipenv. (#5)
Updated all of the testing to use setup.py over pipenv. One of the
limitations of pipenv is the lock file does not support several versions
of python at the same time.

The other consideration is that git_wrapper is a library and thus uses a
editable deps. Any application/package that uses it should be
controlling the dependency chain not this library.

For more information please see: https://docs.pipenv.org/advanced/#pipfile-vs-setup-py

Signed-off-by: Jason Joyce <fuzzball81@gmail.com>