Add Tox & Travis support for py37
Py37 appears to be a version update that broke the most things in a
while, might as well be prepared for the damage.

Travis doesn't currently support it out of the box
(cf. https://github.com/travis-ci/travis-ci/issues/9815 ) so it needs
to be included in the matrix rather than simply added to the Python
list.