Change how we generate version for test pypi builds.
The tooling by default appends '+{hash}' to non-tagged builds,
which is not compliant with the format required by pypi. This
patch chops off the '+' bit, so you get a version like '0.2.9.dev5'
where 'dev5' is the number of commits from the last tag. Provided
github action puts the env var where the python tooling expects it,
this should solve the issue for test deploys (production is unchanged).

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>