Added Prune Functionality To Fetch
Prune related tags were added to the fetch
functions in order to give more functionality
to consumers. This will allow for branches and tags
to be easily pruned simply by calling the fetch function
with the appropriate boolean tags. Added integration
tests to confirm fetch pruning of tags and branches works.